cdk/requirements.txt
aws-cdk-lib==2.64.0
constructs>=10.0.0,<11.0.0
aws-cdk.asset-awscli-v1==2.2.66
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.55
cdk-nag==2.22.5
